Refactor agent service into modular components
